requirements.txt
Django==2.1.2
psycopg2
psycopg2-binary
djangorestframework
requests
mock
